% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proverbs.R
\docType{package}
\name{proverbs}
\alias{proverbs}
\alias{proverbs-package}
\title{proverbs: Print a Daily Bible Proverb to Console.}
\description{
A simple package to grab a Bible proverb corresponding to the day of the month.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bradlindblad/proverbs}
  \item \url{https://bradlindblad.github.io/proverbs/}
  \item Report bugs at \url{https://github.com/bradlindblad/proverbs/issues}
}

}
\author{
\strong{Maintainer}: Brad Lindblad \email{me@bradlindblad.com} [copyright holder]

}
